Revert "Pre-conditioned keys (#2306)"
This reverts commit 8b8a0c5bb3a79cc04e6d6249b2f049e6e25d7aef.
