+++ builder/src/firmware.rs
@@ -331,11 +331,6 @@ pub mod driver_tests {
         ..BASE_FWID
     };
 
-    pub const PRECONDITIONED_KEYS: FwId = FwId {
-        bin_name: "preconditioned_keys",
-        ..BASE_FWID
-    };
-
     pub const PRECONDITIONED_AES: FwId = FwId {
         bin_name: "preconditioned_aes",
         ..BASE_FWID
@@ -629,7 +624,6 @@ pub const REGISTERED_FW: &[&FwId] = &[
     &driver_tests::ML_DSA87_EXTERNAL_MU,
     &driver_tests::ML_KEM,
     &driver_tests::PCRBANK,
-    &driver_tests::PRECONDITIONED_KEYS,
     &driver_tests::PRECONDITIONED_AES,
     &driver_tests::SHA1,
     &driver_tests::SHA256,
