+++ drivers/src/lib.rs
@@ -52,7 +52,6 @@ pub mod pcr_reset;
 mod persistent;
 pub mod pic;
 pub mod preconditioned_aes;
-pub mod preconditioned_key;
 pub mod printer;
 mod sha1;
 mod sha256;
