+++ drivers/test-fw/Cargo.toml
@@ -117,11 +117,6 @@ path = "src/bin/persistent_tests.rs"
 required-features = ["riscv"]
 
 [[bin]]
-name = "preconditioned_keys"
-path = "src/bin/preconditioned_keys_tests.rs"
-required-features = ["riscv"]
-
-[[bin]]
 name = "sha2_512_384acc"
 path = "src/bin/sha2_512_384acc_tests.rs"
 required-features = ["riscv"]
