+++ drivers/test-fw/scripts/vector_gen/src/hmac384_vector_gen.rs
@@ -13,8 +13,7 @@ Abstract:
 --*/
 
 use caliptra_test::crypto::derive_ecdsa_keypair;
-
-use crate::utils::*;
+use openssl::{hash::MessageDigest, pkey::PKey, sign::Signer};
 
 pub struct Hmac384Vector {
     pub seed: [u8; 48],
@@ -36,6 +35,17 @@ impl Default for Hmac384Vector {
     }
 }
 
+fn rand_bytes(buf: &mut [u8]) {
+    openssl::rand::rand_bytes(buf).unwrap()
+}
+
+fn hmac(key: &[u8], msg: &[u8], tag: &mut [u8]) {
+    let pkey = PKey::hmac(key).unwrap();
+    let mut signer = Signer::new(MessageDigest::sha384(), &pkey).unwrap();
+    signer.update(msg).unwrap();
+    signer.sign(tag).unwrap();
+}
+
 pub fn gen_vector(data_len: usize) -> Hmac384Vector {
     let mut vec = Hmac384Vector::default();
 
@@ -47,7 +57,7 @@ pub fn gen_vector(data_len: usize) -> Hmac384Vector {
     let (key, _, _) = derive_ecdsa_keypair(&vec.seed);
 
     let mut out_0 = [0u8; 48];
-    hmac(&key, &vec.data[..], &mut out_0, Digest::SHA384);
+    hmac(&key, &vec.data[..], &mut out_0);
     (_, vec.out_pub_x, vec.out_pub_y) = derive_ecdsa_keypair(&out_0);
 
     vec
