+++ drivers/test-fw/scripts/vector_gen/src/main.rs
@@ -15,13 +15,10 @@ Abstract:
 mod hmac384_kdf_vector_gen;
 mod hmac384_vector_gen;
 mod mdk;
-mod preconditioned_key_extract_gen;
-mod utils;
 
 use crate::hmac384_kdf_vector_gen::Hmac384KdfVector;
 use crate::hmac384_vector_gen::Hmac384Vector;
 use crate::mdk::Mdk;
-use crate::preconditioned_key_extract_gen::PreconditionedKeyExtractVector;
 
 fn hex_arr(bytes: &[u8]) -> String {
     format!(
@@ -84,16 +81,4 @@ fn main() {
     println!("  let mdk = {};", hex_arr(&mdk.mdk));
     println!("  let mek = {};", hex_arr(&mdk::PLAINTEXT_MEK));
     println!("  let encrypted_mek = {};", hex_arr(&mdk.encrypted_mek));
-
-    let vec: PreconditionedKeyExtractVector =
-        preconditioned_key_extract_gen::gen_vector(48, 48, 32);
-    println!("preconditioned_key_extract input_key_length=48");
-    println!("  let input_key = {};", hex_arr(&vec.input_key));
-    println!("  let salt = {};", hex_arr(&vec.input_salt));
-    println!("  let kdf_label = {};", hex_arr(&vec.input_kdf_label));
-    println!("  let checksum = {};", hex_arr(&vec.checksum));
-    println!("  let kdf_output = {};", hex_arr(&vec.kdf_output));
-    println!("  let output_key = {};", hex_arr(&vec.output_key));
-    println!("  let fingerprint = {};", hex_arr(&vec.fingerprint));
-    println!();
 }
