+++ error/src/lib.rs
@@ -2445,12 +2445,6 @@ impl CaliptraError {
             0xa004_0011,
             "Driver Error: AES tag is invalid"
         ),
-        // Skip to leave more error codes for the AES driver.
-        (
-            RUNTIME_DRIVER_PRECONDITIONED_KEY_INVALID_INPUT,
-            0xa004_0020,
-            "Driver Error: preconditioned key usage was invalid"
-        ),
         (
             RUNTIME_DRIVER_PRECONDITIONED_AES_ENCRYPT_ERROR,
             0xa004_0021,
